Team_Projects/SpacePugs/LLMs_to_read_new_laws/pyproject.toml
[project]
name = "hackathon-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "markitdown[all]>=0.1.4",
    "monsterui>=1.0.32",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic-ai>=1.26.0",
    "python-fasthtml>=0.12.35",
]
